Reorganize reexporting of custom derive macros
This change reorganize where and how we reexport our proc macros. Each
custom derive is now reexported with the corresponding trait, so that
if you import for example `diesel::deserialized::Queryable` you will
now import both, the trait and the derive macro. Additionally
documentation for each proc macro is added.